requirements.txt
google-generativeai
gradio
pillow
pandas
lxml
openpyxl
html5lib
beautifulsoup4